--- a/server/v2/stf/branch/bench_test.go
+++ b/server/v2/stf/branch/bench_test.go
@@ -1,6 +1,7 @@
 package branch
 
 import (
+	"encoding/binary"
 	"fmt"
 	"testing"
 
@@ -14,28 +15,60 @@ var (
 )
 
 func Benchmark_CacheStack_Set(b *testing.B) {
+	var sink any
 	for _, stackSize := range stackSizes {
 		b.Run(fmt.Sprintf("StackSize%d", stackSize), func(b *testing.B) {
 			bs := makeBranchStack(b, stackSize)
 			b.ResetTimer()
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
-				_ = bs.Set([]byte{0}, []byte{0})
+				sink = bs.Set([]byte{0}, []byte{0})
 			}
 		})
+	}
+	if sink != nil {
+		b.Fatal("prevent compiler optimization")
 	}
 }
 
 func Benchmark_Get(b *testing.B) {
+	var sink any
 	for _, stackSize := range stackSizes {
 		b.Run(fmt.Sprintf("StackSize%d", stackSize), func(b *testing.B) {
 			bs := makeBranchStack(b, stackSize)
 			b.ResetTimer()
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
-				_, _ = bs.Get([]byte{0})
+				sink, _ = bs.Get([]byte{0})
 			}
 		})
+	}
+	if sink == nil {
+		b.Fatal("prevent compiler optimization")
+	}
+}
+
+func Benchmark_GetSparse(b *testing.B) {
+	var sink any
+	for _, stackSize := range stackSizes {
+		b.Run(fmt.Sprintf("StackSize%d", stackSize), func(b *testing.B) {
+			bs := makeBranchStack(b, stackSize)
+			keys := func() [][]byte {
+				var keys [][]byte
+				for i := 0; i < b.N; i++ {
+					keys = append(keys, numToBytes(i))
+				}
+				return keys
+			}()
+			b.ResetTimer()
+			b.ReportAllocs()
+			for _, key := range keys {
+				sink, _ = bs.Get(key)
+			}
+		})
+	}
+	if sink == nil {
+		b.Fatal("prevent compiler optimization")
 	}
 }
 
@@ -71,7 +104,7 @@ func makeBranchStack(b *testing.B, stackSize int) Store[store.KVStore] {
 		branch = NewStore[store.KVStore](branch)
 		for j := 0; j < elemsInStack; j++ {
 			// create unique keys by including the branch index.
-			key := []byte{byte(i), byte(j)}
+			key := append(numToBytes(i), numToBytes(j)...)
 			value := []byte{byte(j)}
 			err := branch.Set(key, value)
 			if err != nil {
@@ -80,4 +113,8 @@ func makeBranchStack(b *testing.B, stackSize int) Store[store.KVStore] {
 		}
 	}
 	return branch
+}
+
+func numToBytes[T ~int](n T) []byte {
+	return binary.BigEndian.AppendUint64(nil, uint64(n))
 }