--- a/server/v2/stf/branch/bench_test.go
+++ b/server/v2/stf/branch/bench_test.go
@@ -46,6 +46,7 @@ func Benchmark_Get(b *testing.B) {
 	if sink == nil {
 		b.Fatal("benchmark did not run")
 	}
+	sink = nil
 }
 
 func Benchmark_GetSparse(b *testing.B) {