# Description

**Summary of Benchmark Performance Comparison**

---

### Bench Test Type: `_CacheStack_Set`:

This sets the same key in state. Note: Stack size should have no impact here since the top branch is the one where the cache is being set.

- **StackSize1-14**
  - **Previous Performance (storev1.txt):** 48.97 ns/op
  - **Current Performance (old.txt):** 32.93 ns/op
  - **Improvement:** Current code is **1.49× faster**

- **StackSize10-14**
  - **Previous:** 55.87 ns/op
  - **Current:** 46.38 ns/op
  - **Improvement:** Current code is **1.20× faster**

- **StackSize100-14**
  - **Previous:** 56.58 ns/op
  - **Current:** 46.38 ns/op
  - **Improvement:** Current code is **1.22× faster**

---

### Bench Test Type: `_GetCached`

Gets the same key.

NOTE: CacheKV caches the key on the first get, making getting the same key expensive only the first time but in-expensive other times. This works well only when we fetch the same key. GetSparse shows the opposite view.

- **StackSize1-14**
  - **Previous:** 12.56 ns/op
  - **Current:** 25.90 ns/op
  - **Performance Decline:** Current code is **2.06× slower**

- **StackSize10-14**
  - **Previous:** 17.28 ns/op
  - **Current:** 331.1 ns/op
  - **Performance Decline:** Current code is **19.16× slower**

- **StackSize100-14**
  - **Previous:** 16.85 ns/op
  - **Current:** 5,398 ns/op
  - **Performance Decline:** Current code is **320× slower**

---

### Bench Test Type: `_GetSparse`

Fetching always new key from storage.

- **StackSize1-14**
  - **Previous:** 261.8 ns/op
  - **Current:** 22.25 ns/op
  - **Improvement:** Current code is **11.77× faster**

- **StackSize10-14**
  - **Previous:** 1,795 ns/op
  - **Current:** 328.1 ns/op
  - **Improvement:** Current code is **5.47× faster**

- **StackSize100-14**
  - **Previous:** 18,770 ns/op
  - **Current:** 5,292 ns/op
  - **Improvement:** Current code is **3.54× faster**

---

### Bench Test Type: `_Iterate`

- **StackSize1-14**
  - **Previous:** 457.3 ns/op
  - **Current:** 290.4 ns/op
  - **Improvement:** Current code is **1.58× faster**

- **StackSize10-14**
  - **Previous:** 34,090 ns/op
  - **Current:** 13,670 ns/op
  - **Improvement:** Current code is **2.49× faster**

- **StackSize100-14**
  - **Previous:** 47,370,000 ns/op
  - **Current:** 1,441,000 ns/op
  - **Improvement:** Current code is **32.86× faster**

---

### Overall Performance

- **Geometric Mean Time**
  - **Previous:** 626.9 ns/op
  - **Current:** 547.4 ns/op
  - **Overall Improvement:** Current code is approximately **1.15× faster**

---

### Memory Usage

**Bytes per Operation (B/op):**

- **`_CacheStack_Set`**
  - **Previous:** 34 B/op
  - **Current:** 2 B/op
  - **Improvement:** Current code uses **17× less memory**

- **`_GetCached`**
  - **Both Versions:** 1 B/op
  - **Note:** No significant change

- **`_GetSparse`**
  - **Previous:** Ranged from 115 B/op to 13.23 KiB/op
  - **Current:** 0 B/op
  - **Improvement:** Current code eliminates memory allocations in this test

- **`_Iterate`**
  - **Previous vs. Current:** Slight reductions in memory usage with the current code

---

### Allocations per Operation (allocs/op):

- **`_CacheStack_Set`**
  - **Previous:** 3 allocs/op
  - **Current:** 2 allocs/op
  - **Improvement:** Fewer allocations with current code

- **`_GetSparse`**
  - **Previous:** Up to 103 allocs/op
  - **Current:** 0 allocs/op
  - **Improvement:** Current code eliminates allocations

- **`_Iterate`**
  - **Slight Reduction:** Current code has fewer allocations per operation

---

**Conclusion:**

- **Performance Gains:** The current code shows significant performance improvements in the `_CacheStack_Set`, `_GetSparse`, and `_Iterate` benchmarks, especially with larger stack sizes.
- **Performance Losses:** There is a notable performance decline in the `_GetCached` benchmarks, where the current code is slower than the previous version.
- **Memory Efficiency:** The current code generally uses less memory and has fewer allocations per operation, which may contribute to performance improvements in certain benchmarks.

---

This summary provides a detailed comparison of the benchmark tests, highlighting where the current code has improved or declined in performance relative to the previous version.

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review. -->

---

## Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

I have...

* [ ] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title, you can find examples of the prefixes below:
    <!-- * `feat`: A new feature
    * `fix`: A bug fix
    * `docs`: Documentation only changes
    * `style`: Changes that do not affect the meaning of the code (white-space, formatting, missing semi-colons, etc)
    * `refactor`: A code change that neither fixes a bug nor adds a feature
    * `perf`: A code change that improves performance
    * `test`: Adding missing tests or correcting existing tests
    * `build`: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
    * `ci`: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
    * `chore`: Other changes that don't modify src or test files
    * `revert`: Reverts a previous commit -->
* [ ] confirmed `!` in the type prefix if API or client breaking change
* [ ] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#pr-targeting))
* [ ] provided a link to the relevant issue or specification
* [ ] reviewed "Files changed" and left comments if necessary
* [ ] included the necessary unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#testing)
* [ ] added a changelog entry to `CHANGELOG.md`
* [ ] updated the relevant documentation or specification, including comments for [documenting Go code](https://blog.golang.org/godoc)
* [ ] confirmed all CI checks have passed

## Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

Please see [Pull Request Reviewer section in the contributing guide](../CONTRIBUTING.md#reviewer) for more information on how to review a pull request.

I have...

* [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
* [ ] confirmed all author checklist items have been addressed
* [ ] reviewed state machine logic, API design and naming, documentation is accurate, tests and test coverage


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
	- Introduced a suite of benchmark tests for the `cachekv` package, measuring performance for setting and retrieving values, as well as iteration over key-value pairs.
	- Added a new function to convert integers to byte slices for enhanced key generation.

- **Bug Fixes**
	- Improved clarity in comments for the `Get` method within the `Store` struct to better reflect the logic flow.

These enhancements aim to improve performance evaluation and code clarity for users interacting with the caching system.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->